Changed folder structure, some small changes to the routing itself which will be further explained in the next commit.
