+++ mobile-angular-capacitor-app/src/app/app.config.ts
@@ -1,5 +1,5 @@
 import { provideRouter } from '@angular/router';
-import { routes } from './app.routes';
+import { routes } from './core/routing/app.routes';
 import { ApplicationConfig, provideZonelessChangeDetection } from '@angular/core';
 
 export const appConfig: ApplicationConfig = {
