+++ mobile-angular-capacitor-app/src/app/core/routing/account-overview.routes.ts
@@ -9,7 +9,7 @@ export const ACCOUNT_OVERVIEW_ROUTES: Routes = [
   {
     path: 'account-overview-dashboard',
     loadComponent: () =>
-      import('./account-overview-dashboard/account-overview-dashboard.component').then(
+      import('../../features/account-overview/account-overview-dashboard/account-overview-dashboard.component').then(
         (m) => m.AccountOverviewDashboardComponent
       ),
   },
